adds test for deployment
